LNK 2019 fixed

+++ totest/DLinkedListNode.cpp
@@ -60,7 +60,7 @@ DLinkedListNode<T*>::~DLinkedListNode()
 }
 
 
-template<class T, size_t N>
+template<class T, int N>
 void DLinkedListNode<T[N]>::Delete()
 {
 	for (int i = 0; i < N; i++)
@@ -68,7 +68,7 @@ void DLinkedListNode<T[N]>::Delete()
 		data[i] = 0;
 	}
 }
-template<class T, size_t N>
+template<class T, int N>
 DLinkedListNode<T[N]>::DLinkedListNode() : nextElement(nullptr), previousElement(nullptr)
 {
 	for (int i = 0; i < N; i++)
@@ -78,7 +78,7 @@ DLinkedListNode<T[N]>::DLinkedListNode() : nextElement(nullptr), previousElement
 }
 
 
-template<class T, size_t N>
+template<class T, int N>
 DLinkedListNode<T[N]>::DLinkedListNode(T data[N]) : nextElement(nullptr), previousElement(nullptr)
 {
 	for (int i = 0; i < N; i++)
@@ -88,7 +88,7 @@ DLinkedListNode<T[N]>::DLinkedListNode(T data[N]) : nextElement(nullptr), previo
 	}
 }
 
-template<class T, size_t N>
+template<class T, int N>
 DLinkedListNode<T[N]>::~DLinkedListNode()
 {
